+++ soane/tools/clui.py
@@ -6,6 +6,30 @@ import click
 
 from soane import tools
 
+class AbbrGroup(click.Group):
+    '''
+    A custom Group type that supports abbreviated commands.
+    '''
+
+    def get_command(self, ctx, comm_name):
+        '''
+        Return a Command from a complete or abbreviated name.
+        '''
+
+        names = [
+            name for name in self.list_commands(ctx)
+            if name.startswith(comm_name)
+        ]
+
+        if not names:
+            return None
+        elif len(names) == 1:
+            return super().get_command(ctx, names[0])
+        else:
+            ambi = ', '.join(sorted(names))
+            ctx.fail(f'Ambiguous abbreviation, did you mean: {ambi}?')
+
+
 class BookType(click.ParamType):
     '''
     A custom parameter type for Book paths.
@@ -23,6 +47,7 @@ class BookType(click.ParamType):
             self.fail(f'Book path {path!r} does not exist')
 
         return path
+
 
 class NameType(click.ParamType):
     '''