+++ tests/test_tools/test_clui.py
@@ -9,6 +9,27 @@ import pytest
 
 from soane.tools import clui
 
+def test_AbbrGroup():
+    # setup
+    @click.group(cls=clui.AbbrGroup)
+    def group():
+        pass
+
+    @group.command()
+    def test_one():
+        pass
+
+    @group.command()
+    def test_two():
+        pass
+
+    # success
+    assert group.get_command(None, 'test-o')   == test_one
+    assert group.get_command(None, 'test-one') == test_one
+
+    with pytest.raises(click.exceptions.UsageError):
+        group.get_command(click.Context(group), 'test')
+
 def test_BookType(tmpdir):
     # setup
     path = tmpdir.join('book.ext')